+++ CHANGES.rst
@@ -10,6 +10,9 @@
 
 - Add support for PyPy3.
 
+- Add interfaces for all schema fields defined in
+  ``nti.contentfragments.schema`` and make the respective classes
+  implement them.
 
 1.1.1 (2018-06-29)
 ==================